+++ index.mjs
@@ -307,8 +307,6 @@ var methods = {
       case 'Tab':
         e.preventDefault();
 
-        if(selectedValue && selectedValue[optionIdentifier] === items[hoverItemIndex][optionIdentifier]) return;
-
         this.set({activeItemIndex: hoverItemIndex});
         this.handleSelect(items[hoverItemIndex]);
         break;
@@ -1256,9 +1254,14 @@ function ondestroy$1() {
 function onstate({changed, current, previous}) {
   if (!previous) return;
 
-  if (changed.selectedValue && current.selectedValue) {
-    if (!previous.selectedValue || current.selectedValue[current.optionIdentifier] != previous.selectedValue[current.optionIdentifier])
+  if (!current.isMulti && changed.selectedValue && current.selectedValue) {       
+    if (!previous.selectedValue || JSON.stringify(current.selectedValue[current.optionIdentifier]) != JSON.stringify(previous.selectedValue[current.optionIdentifier])) {
       this.fire('select', current.selectedValue);
+    }
+  }
+
+  if (current.isMulti && JSON.stringify(current.selectedValue) != JSON.stringify(previous.selectedValue)) {
+    this.fire('select', current.selectedValue);
   }
 
   if (changed.listOpen) {