+++ test/public/bundle.js
@@ -24180,7 +24180,7 @@
 		Object.defineProperty(exports, '__esModule', { value: true });
 
 	})));
-
+	//# sourceMappingURL=svelte.js.map
 	});
 
 	unwrapExports(svelte);
@@ -24747,8 +24747,6 @@
 	        break;
 	      case 'Tab':
 	        e.preventDefault();
-
-	        if(selectedValue && selectedValue[optionIdentifier] === items[hoverItemIndex][optionIdentifier]) return;
 
 	        this.set({activeItemIndex: hoverItemIndex});
 	        this.handleSelect(items[hoverItemIndex]);
@@ -25697,9 +25695,14 @@
 	function onstate({changed, current, previous}) {
 	  if (!previous) return;
 
-	  if (changed.selectedValue && current.selectedValue) {
-	    if (!previous.selectedValue || current.selectedValue[current.optionIdentifier] != previous.selectedValue[current.optionIdentifier])
+	  if (!current.isMulti && changed.selectedValue && current.selectedValue) {       
+	    if (!previous.selectedValue || JSON.stringify(current.selectedValue[current.optionIdentifier]) != JSON.stringify(previous.selectedValue[current.optionIdentifier])) {
 	      this.fire('select', current.selectedValue);
+	    }
+	  }
+
+	  if (current.isMulti && JSON.stringify(current.selectedValue) != JSON.stringify(previous.selectedValue)) {
+	    this.fire('select', current.selectedValue);
 	  }
 
 	  if (changed.listOpen) {
@@ -27263,6 +27266,7 @@
 	        });
 	    });
 	}
+	//# sourceMappingURL=tape-modern.esm.js.map
 
 	// setup
 	const target = document.querySelector('main');
@@ -29087,6 +29091,39 @@
 	  select.set({selectedValue: {value: 'cake', label: 'Cake'}});
 
 	  t.ok(!item);
+	  listener.cancel();
+	  select.destroy();
+	});
+
+	test('when isMulti and item is selected or state changes then check selectedValue[optionIdentifier] has changed before firing "select" event', async (t) => {
+	  const div = document.createElement('div');
+	  document.body.appendChild(div);
+
+	  const select = new Select({
+	    target,
+	    data: {
+	      isMulti: true,
+	      items,
+	      selectedValue: [
+	        {value: 'pizza', label: 'Pizza'},
+	        {value: 'chips', label: 'Chips'},
+	      ],
+	    }
+	  });
+
+	  let item = undefined;
+
+	  const listener = select.on('select', () => {
+	    item = true;
+	  });
+
+	  select.set({selectedValue: [{value: 'pizza', label: 'Pizza'},{value: 'chips', label: 'Chips'}]});
+	  t.ok(!item);
+	  item = false;
+
+	  select.set({selectedValue: [{value: 'pizza', label: 'Pizza'}]});
+	  t.ok(item);
+
 	  listener.cancel();
 	  select.destroy();
 	});